Semi-major update:
- Player -> Color (to eliminate confusing semantics)
- Added CPM
- Update PGN parsing
- Add PieceType type
- Add CastleSide type
- Fix a couple of variable namings
- Added a zobrist hash test
- Expanded HashKey a bit to be more flexible

+++ src/Rudzoft.ChessLib/Types/File.cs
@@ -167,7 +167,7 @@ public readonly record struct File(Files Value)
     public static bool operator false(File f) => !f.IsOk;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static implicit operator int(File f) => (int)f.Value;
+    public static implicit operator byte(File f) => (byte)f.Value;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public int AsInt() => (int)Value;
